UI - Pause Menu
Added a working pause menu that can save user settings, the same as the main menu. Completing this means that we can navigate from the main menu, to a level, to another level, back to the main menu and anywhere in between. *Onlyworks in viewport or New Editor Window, otherwise the levels don't have character movement*